backend/requirements.txt
apng==0.3.4
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
distro==1.9.0
Django==5.0.6
djangorestframework==3.15.1
drf-extra-fields==3.7.0
exceptiongroup==1.2.0
ExifRead==3.0.0
filetype==1.2.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
intervaltree==3.1.0
numpy==1.26.4
outcome==1.3.0.post0
peewee==3.17.0
pillow==10.2.0
pipenv==2023.11.15
prettytable==3.10.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
pypng==0.20220715.0
PySocks==1.7.1
python-dotenv==1.0.1
python-magic==0.4.27
python-slugify==8.0.1
python-xmp-toolkit==2.0.2
pytz==2024.1
requests==2.31.0
selenium==4.16.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
stegoveritas==1.10
stegoveritas-binwalk==2.1.2
stegoveritas-pfp==0.2.3
stegoveritas-py010parser==0.1.10
text-unidecode==1.3
trio==0.23.2
trio-websocket==0.11.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.1.0
wcwidth==0.2.13
wsproto==1.2.0
zope.interface==6.4.post2
